This commit refactors the monolithic `main.py` file into a more organized and maintainable modular structure. The application's logic has been split into functional blocks, each residing in its own file under a new `api/` directory.
The new structure is as follows:
- `main.py`: The main application entry point, responsible for initializing the FastAPI app and including the API routers.
- `api/models.py`: Contains all Pydantic models used for request and response validation.
- `api/helpers.py`: Contains helper functions, such as for parsing request parameters.
- `api/computation.py`: Contains the core astronomical calculation logic.
- `api/background.py`: Contains functions for triggering and managing background precomputation tasks.
- `api/routes/`: A directory containing the API routers for different parts of the application (e.g., `celestial.py`, `asteroids.py`, `comets.py`).

This refactoring improves the overall code quality, making it easier to navigate, understand, and maintain.
